+++ src/timezone.js
@@ -1,8 +1,7 @@
 const TIMEZONE_REG = /Z|[A-Z]{3}|[+-]\d{2}:?\d{2}$/;
-const ISO_DATE_REG = /^[+-]?\d{4,5}(-?\d{2})?(-?\d{2})?$/;
 
 export function isAmbiguousTimeZone(str) {
-  return !TIMEZONE_REG.test(str) && !ISO_DATE_REG.test(str);
+  return !TIMEZONE_REG.test(str);
 }
 
 // This method returns the offset for a given timezone,